move getAccounts method
