+++ utils/keys.go
@@ -18,6 +18,26 @@ type Account struct {
 	Delegations []stakingtypes.Delegation `json:"delegations"`
 }
 
+// GetAccounts is a method to retrieve the binaries keys from the configured
+// keyring backend and stores it in the Binary struct.
+func (bin Binary) GetAccounts() error {
+	out, err := ExecuteBinaryCmd(&bin, BinaryCmdArgs{
+		Subcommand: []string{"keys", "list", "--output=json"},
+	})
+	if err != nil {
+		return err
+	}
+
+	accounts, err := ParseAccountsFromOut(out)
+	if err != nil {
+		return err
+	}
+
+	bin.Accounts = accounts
+
+	return nil
+}
+
 // FilterAccountsWithDelegations filters the given list of accounts for those, which are used for staking.
 func FilterAccountsWithDelegations(bin *Binary) ([]Account, error) {
 	var stakingAccs []Account
